+++ api-types/src/shaders.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default, facet::Facet)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default, facet::Facet, specta::Type)]
 pub struct ShaderFilterFacets {
     pub query: Option<String>,
     pub loaders: Option<Vec<(String, String)>>,
@@ -10,7 +10,7 @@ pub struct ShaderFilterFacets {
     pub game_versions: Option<Vec<String>>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct ShaderPack {
     pub id: String,
     pub name: String,
@@ -19,17 +19,17 @@ pub struct ShaderPack {
     pub description: Option<String>,
     pub file_path: String,
     pub file_name: String,
-    pub file_size: u64,
+    pub file_size: u32,
     pub compatible_versions: Vec<String>,
     pub enabled: bool,
     pub source_url: Option<String>,
     pub thumbnail: Option<String>,
     pub shader_loader: ShaderLoader,
-    pub installed_date: i64,
-    pub last_used: Option<i64>,
+    pub installed_date: i32,
+    pub last_used: Option<i32>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 #[facet(rename_all = "snake_case")]
 #[serde(rename_all = "snake_case")]
 #[repr(u8)]
@@ -40,7 +40,7 @@ pub enum ShaderLoader {
     Vanilla,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct ShaderSettings {
     pub quality: ShaderQuality,
     pub shadows: bool,
@@ -51,7 +51,7 @@ pub struct ShaderSettings {
     pub custom_settings: HashMap<String, String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 #[facet(rename_all = "snake_case")]
 #[serde(rename_all = "snake_case")]
 #[repr(u8)]
@@ -63,7 +63,7 @@ pub enum ShaderQuality {
     Custom,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct ShaderDownload {
     pub id: String,
     pub name: String,
@@ -77,12 +77,12 @@ pub struct ShaderDownload {
     pub minecraft_versions: Vec<String>,
     pub shader_loader: ShaderLoader,
     pub rating: f32,
-    pub downloads: u64,
-    pub size_mb: u64,
+    pub downloads: u32,
+    pub size_mb: u32,
     pub source: ShaderSource,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 #[facet(rename_all = "snake_case")]
 #[serde(rename_all = "snake_case")]
 #[repr(u8)]
