+++ api-types/src/skins.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-#[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Default, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Default, facet::Facet, specta::Type)]
 #[facet(rename_all = "snake_case")]
 #[serde(rename_all = "snake_case")]
 #[repr(u8)]
@@ -28,37 +28,37 @@ impl SkinModel {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, facet::Facet)]
+#[derive(Debug, Clone, Serialize, Deserialize, facet::Facet, specta::Type)]
 pub struct SkinUploadConfig {
     pub model: SkinModel,
     pub file_path: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, facet::Facet, specta::Type)]
 pub struct SkinUploadResponse {
     pub success: bool,
     pub message: String,
     pub model_used: SkinModel,
 }
 
-#[derive(Debug, Serialize, Deserialize, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, facet::Facet, specta::Type)]
 pub struct CurrentSkin {
     pub model: SkinModel,
     pub url: Option<String>,
     pub has_skin: bool,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct AccountSkin {
     pub id: String,
     pub name: String,
     pub url: Option<String>,
     pub model: SkinModel,
     pub is_current: bool,
-    pub uploaded_date: Option<i64>,
+    pub uploaded_date: Option<i32>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct AccountCape {
     pub id: String,
     pub state: String,
@@ -66,7 +66,7 @@ pub struct AccountCape {
     pub alias: Option<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct PlayerProfile {
     pub id: String,
     pub name: String,
@@ -74,14 +74,14 @@ pub struct PlayerProfile {
     pub capes: Vec<AccountCape>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct CustomSkinsRoot {
     #[serde(rename = "customSkins")]
     pub custom_skins: HashMap<String, CustomSkinEntry>,
     pub version: Option<u32>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet)]
+#[derive(Debug, Serialize, Deserialize, Clone, facet::Facet, specta::Type)]
 pub struct CustomSkinEntry {
     #[serde(default, rename = "capeId")]
     pub cape_id: String,
